+++ BaiTap1/BaiTap1/Controllers/StudentController.cs
@@ -1,4 +1,7 @@
-﻿using BaiTap1.Models;
+﻿using BaiTap1.DTO.Course;
+using BaiTap1.DTO.Student;
+using BaiTap1.Mapper;
+using BaiTap1.Models;
 using BaiTap1.Services;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -14,53 +17,51 @@ namespace BaiTap1.Controllers
         {
             _studentService = studentService;
         }
-
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Student>>> GetAllStudents()
+        public async Task<IActionResult> GetAllStudents()
         {
             var students = await _studentService.GetAllStudentAsync();
             return Ok(students);
         }
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Student>> GetStudentById(int id)
-        {
-            var students = await _studentService.GetStudentByIdAsync(id);
-            if (students == null) 
-            { 
-                return NotFound();
-            }
-            return Ok(students);
-        }
-        [HttpPost]
-        public async Task<ActionResult<Student>> CreateStudent([FromBody] Student student)
-        {
-            var createStudent = await _studentService.CreateStudentAsync(student);
-            return CreatedAtAction(nameof(GetStudentById), new { id = createStudent.ID }, createStudent);
-        }
-        [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateStudent(int id, [FromBody] Student student)
-        {
-            if (id != student.ID)
-            {
-                return BadRequest("ID sinh viên không khớp.");
-            }
-            var updateStudent = await _studentService.UpdateStudentByIdAsync(id, student);
-            if(updateStudent == null)
-            {
-                return NotFound();
-            }
-            return NoContent();
-        }
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteStudent(int id)
-        {
-            var result = await _studentService.DeleteStudentByIdAsync(id);
-            if (!result)
-            {
-                return NotFound();
-            }
 
-            return NoContent();
+        [HttpGet("{id}")]
+        public async Task<IActionResult> GetByIdStudent(int id)
+        {
+            var student = await _studentService.GetStudentByIdAsync(id);
+            if (student == null)
+            {
+                return NotFound();
+            }
+            return Ok(student);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> CreateStudent(CreateStudentDTO createStudentDTO)
+        {
+            var createdStudent = await _studentService.CreateStudentAsync(createStudentDTO);
+            return CreatedAtAction(nameof(GetByIdStudent), new { id = createdStudent.Id }, createdStudent);
+        }
+
+        [HttpPut("{id}")]
+        public async Task<IActionResult> UpdateStudent(int id, UpdateStudentDTO updateStudentDTO)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(new ResponAPI { Id = 1, Description = "Dữ liệu không hợp lệ", Data = null });
+            }
+            var updatedCourse = await _studentService.UpdateStudentByIdAsync(id, updateStudentDTO);
+            return Ok(updateStudentDTO);
+        }
+
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteCourse(int id)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(new ResponAPI { Id = 1, Description = "Dữ liệu không hợp lệ", Data = null });
+            }
+            var result = await _studentService.DeleteStudentByIdAsync(id);
+            return Ok(result);
         }
     }
 }
